requirements.txt
splitcopy==1.1.6
prompt_toolkit==3.0.18
junos_eznc==2.5.4
lxml==4.6.2
dataclasses==0.6
PyYAML==5.4.1
jsnapy==1.3.5
